fix: validate release workflows

+++ src/gauge.cpp
@@ -3,6 +3,8 @@
 #include "../include/engine_sim_application.h"
 #include "../include/constants.h"
 
+#include <climits>
+
 Gauge::Gauge() {
     m_thetaMin = (float)constants::pi;
     m_thetaMax = 0.0f;
@@ -49,7 +51,7 @@ void Gauge::destroy() {
 
 void Gauge::update(float dt) {
     const float value = std::fmaxf(m_min, std::fmin(m_max, m_value));
-    const float needle_s = std::powf((value - m_min) / std::abs(m_max - m_min), m_gamma);
+    const float needle_s = std::pow((value - m_min) / std::abs(m_max - m_min), m_gamma);
     const float F =
         m_needleKs * (needle_s - m_needlePosition)
         - m_needleKd * m_needleVelocity;
@@ -76,7 +78,7 @@ void Gauge::render() {
     GeometryGenerator::Line2dParameters lineParams;
     generator->startShape();
     for (int i = 0; i <= std::abs(m_max - m_min); i += m_minorStep) {
-        const float s = std::powf(i / std::abs(m_max - m_min), m_gamma);
+        const float s = std::pow(i / std::abs(m_max - m_min), m_gamma);
         const float theta = s * m_thetaMax + (1 - s) * m_thetaMin;
 
         const float tickLength = (i % m_majorStep) == 0
@@ -139,10 +141,10 @@ void Gauge::render() {
         ringParams.outerRadius = outerRadius + pixelsToUnits(band.radial_offset);
         ringParams.innerRadius = outerRadius + pixelsToUnits(band.radial_offset - band.width);
 
-        const float s0 = std::powf((band.start - m_min) / std::abs(m_max - m_min), m_gamma);
+        const float s0 = std::pow((band.start - m_min) / std::abs(m_max - m_min), m_gamma);
         const float angle0 = s0 * m_thetaMax + (1 - s0) * m_thetaMin;
 
-        const float s1 = std::powf((band.end - m_min) / std::abs(m_max - m_min), m_gamma);
+        const float s1 = std::pow((band.end - m_min) / std::abs(m_max - m_min), m_gamma);
         const float angle1 = s1 * m_thetaMax + (1 - s1) * m_thetaMin;
 
         ringParams.startAngle = std::fminf(angle0, angle1) + band.shorten_end;
